Limit device polling where data is always available, update keyboard
