+++ toon/input/keyboard.py
@@ -1,44 +1,40 @@
 import numpy as np
-import keyboard as kb
-from toon.input.base_input import DummyTime
+from toon.input.base_input import BaseInput, DummyTime
 
-
-class Keyboard(object):
+class Keyboard(BaseInput):
     def __init__(self,
                  clock_source=DummyTime,
-                 keys=None):
-
+                 keys=None,
+                 multiprocess=False,
+                 nrow=10):
         if keys is None:
             raise ValueError('`keys` must be a list of keys of interest.')
 
-        self.time = clock_source
+        BaseInput.__init__(self, clock_source, multiprocess, (10, len(keys)))
         self._lenkeys = len(keys)
         self._keys = keys
         self._buffer = np.full(self._lenkeys, 0)
-        self._outbuffer = np.full(self._lenkeys, 0)
-        self.name = type(self).__name__
+        self._sampling_period = 0.01
 
-    def __enter__(self):
-        self._start_time = self.time.getTime()
+    def _init_device(self):
+        import keyboard
+        self._device = keyboard
         self._buffer[:] = 0
-        self._outbuffer[:] = 0
         n = 0
         for key in self._keys:
-            kb.add_hotkey(key, self._add_array, (n,),
-                          timeout=0.001)
-            kb.add_hotkey(key, self._rem_array, (n,),
-                          timeout=0.001,
-                          trigger_on_release=True)
+            keyboard.add_hotkey(key, self._add_array, (n,), timeout=0)
+            keyboard.add_hotkey(key, self._rem_array, (n,), timeout = 0, trigger_on_release=True)
             n += 1
-        return self
 
-    def __exit__(self, type, value, traceback):
-        kb.clear_all_hotkeys()
 
-    def read(self):
-        timestamp = self.time.getTime()
-        np.copyto(self._outbuffer, self._buffer)
-        return self._outbuffer, timestamp
+    def _read(self):
+        return self._buffer, self.time.getTime()
+
+    def _stop_device(self):
+        self._device.clear_all_hotkeys()
+
+    def _close_device(self):
+        pass
 
     def _add_array(self, index):
         self._buffer[index] = 1
